Program is just about ready for graphing. Technically stuff is being graphed right now but it needs to be prettier.
