+++ FileConversion.py
@@ -11,8 +11,8 @@ def FileConversion(cvp,xlp):
         reader = csv.reader(f)
         for row in reader:
             ws.append(row)
-
     wb.save(xlp)
+    
     ws.delete_cols(1)
     ws.delete_cols(2,2)
     ws.delete_cols(3,2)
@@ -25,5 +25,6 @@ def FileConversion(cvp,xlp):
     for cell in ws['A']:
         cell.value = cell.value[:-7]
         wb.save(xlp)
+
 
 # FileConversion(csvpath,path)
