+++ Graph.py
@@ -1,26 +1,23 @@
-import ScrapeExcel, random
+import random, consolidate, openpyxl, path
 
+from openpyxl import load_workbook
 import matplotlib.pyplot as plt
 import numpy as np
-import ScrapeExcel as data
+
+path = path.path
+wb = load_workbook(path)
+ws = wb.active
 
 # variables
-dict = ScrapeExcel.dict
-x = []
-y = []
+dict = consolidate.vol_dict
 linecolor = ['r', 'g', 'b', 'c', 'm', 'y', 'k', 'w']
 line = ['-', '--', '-.']
 
-def getallkeys(dict):
-    for key, value in dict.items():
-        yield key
-    if isinstance(value, dict):
-        yield from getallkeys(value)
+consolidate.Consolidate(ws,path)
 
 
-for x in getallkeys(dict):
-    # x.append(key)
-    # y.append(5)
-    plt.plot(x, random.randint(200,500), f'o{random.choice(line)}{random.choice(linecolor)}')
+for key, total in dict.items():
+        print(f"Date: {key[0]}, Title: {key[1]}, Sum: {total}")
+        plt.plot(key[0], total, f'o{random.choice(line)}{random.choice(linecolor)}')
 
 plt.show()
