+++ products.py
@@ -6,15 +6,11 @@ path = path.path
 wb = load_workbook(path)
 ws = wb.active
 
-
-def Products(ws):
-    for row in ws.iter_rows():
-        weight = 0
-        reps = 0
+def Products(wsheet,xlp):
+    currentrow = 0
+    for row in wsheet.iter_rows(min_col=5, max_col=5):
+        currentrow += 1
         for cell in row:
-            match cell.column_letter:
-                case 'C': weight = cell.value
-                case 'D': reps = cell.value
-        ws['E'] = weight*reps
-        wb.save(path)
-        break
+            wsheet[f"E{currentrow}"] = wsheet[f"C{currentrow}"].value * wsheet[f"D{currentrow}"].value
+
+# Products(ws,path)
