+++ removetreadmill.py
@@ -6,13 +6,13 @@ wb = load_workbook(path.path)
 ws = wb.active
 
 def RemoveTreadmill(wsheet,xlp):
-        currentrow = 0
         for row in wsheet.iter_rows(min_col=2, max_col=2):
-                currentrow += 1
                 for cell in row:
                         if cell.value == "Treadmill":
-                                ws.delete_rows(currentrow)
-                                wb.save(xlp)
-                                currentrow -= 1
+                                ws.delete_rows(cell.row)
                                 continue
+                        wb.save(xlp)
+                                
+                                    
+                        
 # RemoveTreadmill(ws,path.path)
