--- setup.py
@@ -1,10 +1,0 @@
-import openpyxl, datetime, path
-
-from datetime import datetime
-from openpyxl import Workbook
-
-path = path.path
-wb = Workbook('workouts.xlsx')
-ws = wb.active
-
-
